Revert "Revert "Revert "UltimosCambios 07/04"""
This reverts commit ea0d69f3abd7b519a42b9435906985475ace0ca5.

+++ macowins/src/main/java/Grupo4/Tipo.java
@@ -1,6 +1,6 @@
 package Grupo4;
 
 public interface Tipo {
-	double precioBase();
+	float precioBase();
 	
 }
